RSC: Split transform plugin into client and server
